zcash_client_backend: introduce data_api::locking module
Centralize the note-locking vocabulary into a dedicated
zcash_client_backend::data_api::locking module. This is a move-only
change with zero semantic difference; every previous path keeps
resolving via re-exports:

- LockOwner moves from crate::wallet (re-exported at its old path).
- LockError moves from data_api::error (re-exported at its old path).
- LockRequest, unlock_proposal_inputs, lock_proposal_inputs (now
  pub(crate)) and the module-private proposal_input_refs move from
  data_api::wallet (LockRequest and unlock_proposal_inputs re-exported
  at their old paths).
- LockedInputPolicy and LockFilter move from
  data_api::wallet::input_selection (both re-exported at their old
  paths).

OutputRef stays in crate::wallet (proposal validation uses it beyond
locking) and NonEmptyBTreeSet stays in input_selection (it is generic);
the locking module imports both.

The module rustdoc now states the locking semantics once (the
locked/eligible complement, acquisition and no-stealing rules, the
advisory nature of locks, the four release paths, and the expiry race)
and maps every integration point in the codebase by role.

This is an atomic move: it cannot be split into a "copy then remove"
pair, because the moved types appear in the WalletWrite trait method
signatures, so a transient state with both the old and the new type
present would fail to type-check (two distinct types of the same name).
Moving the definition and leaving a re-export in the same commit keeps
a single type identity.

Reviewer reproduction:
- Verify it is a pure move (no logic change) by rendering the diff with
  moved-line detection; every line of the new data_api/locking.rs shows
  up as moved out of error.rs, wallet.rs, input_selection.rs, or
  src/wallet.rs, and the only non-moved additions are the
  `pub mod locking;` line and the `pub use ...locking::{...}` re-exports
  left at each old path:
    git show --color-moved=dimmed-zebra <this commit>
- Confirm the old public paths still resolve (the API is unchanged):
    cargo build --release -p zcash_client_backend --all-features
- Confirm behavior is unchanged:
    cargo test --release -p zcash_client_backend --all-features
