+++ zcash_client_backend/src/data_api.rs
@@ -136,6 +136,7 @@ pub mod chain;
 pub mod defaults;
 pub mod error;
 pub mod ll;
+pub mod locking;
 pub mod scanning;
 pub mod wallet;
