+++ zcash_client_backend/src/data_api/wallet.rs
@@ -44,11 +44,12 @@ use std::{
 use shardtree::error::{QueryError, ShardTreeError};
 
 use super::InputSource;
+use super::locking::lock_proposal_inputs;
+pub use super::locking::{LockRequest, unlock_proposal_inputs};
 use crate::{
     data_api::{
         Account, MaxSpendMode, NoteCommitmentTree, SentTransaction, SentTransactionOutput,
-        WalletCommitmentTrees, WalletRead, WalletWrite,
-        error::{Error, LockError},
+        WalletCommitmentTrees, WalletRead, WalletWrite, error::Error,
         wallet::input_selection::propose_send_max,
     },
     decrypt_transaction,
@@ -56,7 +57,7 @@ use crate::{
         ChangeStrategy, DustOutputPolicy, StandardFeeRule, standard::SingleOutputChangeStrategy,
     },
     proposal::{Proposal, ProposalError, Step, StepOutputIndex},
-    wallet::{LockOwner, Note, OutputRef, OvkPolicy, Recipient},
+    wallet::{Note, OvkPolicy, Recipient},
 };
 use sapling::{
     note_encryption::{PreparedIncomingViewingKey, try_sapling_note_decryption},
@@ -668,119 +669,6 @@ impl ConfirmationsPolicy {
     }
 }
 
-/// Returns the [`OutputRef`] identifying each output that the given proposal consumes as an
-/// input.
-///
-/// Each note or UTXO selected for spending is an *input* to the proposal's transaction, but is at
-/// the same time an *output* of the earlier transaction that created it; an [`OutputRef`] names it
-/// by that creating transaction's id, which is the stable identity the lock tables are keyed on.
-fn proposal_input_refs<FeeRuleT, NoteRef>(
-    proposal: &Proposal<FeeRuleT, NoteRef>,
-) -> Vec<OutputRef> {
-    proposal
-        .steps()
-        .iter()
-        .flat_map(|step| {
-            step.shielded_inputs()
-                .into_iter()
-                .flat_map(|shielded_inputs| {
-                    shielded_inputs.notes().iter().map(|note| {
-                        OutputRef::new(
-                            *note.txid(),
-                            PoolType::Shielded(note.note().pool()),
-                            u32::from(note.output_index()),
-                        )
-                    })
-                })
-                .chain(step.transparent_inputs().iter().map(|utxo| {
-                    let outpoint = utxo.outpoint();
-                    OutputRef::new(
-                        TxId::from_bytes(*outpoint.hash()),
-                        PoolType::TRANSPARENT,
-                        outpoint.n(),
-                    )
-                }))
-        })
-        .collect()
-}
-
-/// A request to lock the inputs selected by a proposal, made when calling one of the
-/// proposal-creation functions ([`propose_transfer`] and friends).
-///
-/// The caller supplies the [`LockOwner`] under which the locks are taken and must retain it: the
-/// owner token is what authorizes releasing the locks with [`unlock_proposal_inputs`], and what
-/// allows the same flow to re-lock its own inputs when retrying after a crash.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct LockRequest {
-    owner: LockOwner,
-    for_blocks: u32,
-}
-
-impl LockRequest {
-    /// Constructs a request to lock the proposal's inputs on behalf of `owner` until
-    /// `for_blocks` blocks past the proposal's target height.
-    ///
-    /// Choose `for_blocks` conservatively with respect to the worst-case time between proposal
-    /// creation and transaction storage: once the lock expires, a concurrent proposal may
-    /// select the same inputs.
-    pub fn new(owner: LockOwner, for_blocks: u32) -> Self {
-        Self { owner, for_blocks }
-    }
-
-    /// Returns the owner under which the locks will be taken.
-    pub fn owner(&self) -> LockOwner {
-        self.owner
-    }
-
-    /// Returns the number of blocks past the proposal's target height at which the locks will
-    /// expire.
-    pub fn for_blocks(&self) -> u32 {
-        self.for_blocks
-    }
-}
-
-/// Locks all inputs selected by the given proposal, preventing them from being
-/// selected by subsequent proposals. The lock expires at the given height.
-#[allow(clippy::type_complexity)]
-fn lock_proposal_inputs<DbT, FeeRuleT, NoteRef, TE, SE, FE, CE>(
-    wallet_db: &mut DbT,
-    proposal: &Proposal<FeeRuleT, NoteRef>,
-    owner: LockOwner,
-    lock_expiry_height: BlockHeight,
-) -> Result<(), Error<DbT::Error, TE, SE, FE, CE, NoteRef>>
-where
-    DbT: WalletWrite,
-{
-    match wallet_db.lock_outputs(&proposal_input_refs(proposal), owner, lock_expiry_height) {
-        Ok(_) => Ok(()),
-        Err(LockError::LockFailure(out_ref)) => {
-            Err(Error::Proposal(ProposalError::InputsLocked(out_ref)))
-        }
-        Err(LockError::Storage(e)) => Err(Error::DataSource(e)),
-    }
-}
-
-/// Unlocks all inputs selected by the given proposal, reversing the locks acquired when the
-/// proposal was created with a [`LockRequest`] under the same `owner`.
-///
-/// This is useful when a proposal is rejected or abandoned after its inputs were locked, so that
-/// the outputs become available for selection and balance computation once again. Because
-/// unlocking is scoped to `owner`, inputs that are not locked, or whose locks are held by a
-/// different owner (for example a concurrently-created proposal), are left unchanged.
-pub fn unlock_proposal_inputs<DbT, FeeRuleT, NoteRef>(
-    wallet_db: &mut DbT,
-    proposal: &Proposal<FeeRuleT, NoteRef>,
-    owner: LockOwner,
-) -> Result<(), DbT::Error>
-where
-    DbT: WalletWrite,
-{
-    for output_ref in proposal_input_refs(proposal) {
-        wallet_db.unlock_output(&output_ref, owner)?;
-    }
-    Ok(())
-}
-
 /// Select transaction inputs, compute fees, and construct a proposal for a transaction or series
 /// of transactions that can then be authorized and made ready for submission to the network with
 /// [`create_proposed_transactions`].
@@ -813,6 +701,7 @@ where
 /// proposal may select and spend the same inputs. Choose the window conservatively with respect
 /// to the worst-case time between proposal creation and transaction storage.
 ///
+/// [`LockOwner`]: crate::wallet::LockOwner
 /// [`WalletWrite::lock_outputs`]: crate::data_api::WalletWrite::lock_outputs
 /// [`WalletWrite::store_transactions_to_be_sent`]: crate::data_api::WalletWrite::store_transactions_to_be_sent
 /// [`WalletWrite::clear_locked_outputs`]: crate::data_api::WalletWrite::clear_locked_outputs
